src/requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
cloudinary==1.7.0
dj-database-url==0.4.2
Django==1.10.6
django-braces==1.11.0
django-comments-xtd==2.0.3
django-contrib-comments==1.8.0
django-model-utils==3.0.0
django-registration==2.2
django-smart-selects==1.5.2
django-star-ratings==0.5.5
djangorestframework==3.6.3
docutils==0.13.1
gunicorn==19.7.1
idna==2.5
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
whitenoise==3.3.0
